// src/app/new-application/interfaces/application.ts
export interface Application {
    university_name: string
    course_level: string
    course_intake: string
    course_name: string
    student_name: string
    student_email: string
    student_dob: string
    visa_refusal: string
    nationality: string
    student_number: string
    passport_number: string
    passport_expire_date: string
    user_id: number
    updated_at: string
    created_at: string
    id: number
    application_id: string
}